Rename npc package to character for better clarity
Moved all classes and references from the 'npc' package to the 'character' package to align with updated naming conventions. This improves semantic clarity and consistency across the codebase.
