+++ plugin/src/main/java/net/thenextlvl/service/controller/character/CitizensCharacterController.java
@@ -3,9 +3,9 @@ package net.thenextlvl.service.controller.character;
 import net.citizensnpcs.api.CitizensAPI;
 import net.citizensnpcs.api.npc.NPC;
 import net.thenextlvl.service.ServicePlugin;
-import net.thenextlvl.service.api.npc.Character;
-import net.thenextlvl.service.api.npc.CharacterCapability;
-import net.thenextlvl.service.api.npc.CharacterController;
+import net.thenextlvl.service.api.character.Character;
+import net.thenextlvl.service.api.character.CharacterCapability;
+import net.thenextlvl.service.api.character.CharacterController;
 import net.thenextlvl.service.model.character.citizens.CitizensCharacter;
 import org.bukkit.Location;
 import org.bukkit.World;
