+++ plugin/src/main/java/net/thenextlvl/service/listener/CitizensListener.java
@@ -5,11 +5,11 @@ import net.citizensnpcs.api.event.NPCDamageByEntityEvent;
 import net.citizensnpcs.api.event.NPCDamageEvent;
 import net.citizensnpcs.api.event.NPCLeftClickEvent;
 import net.citizensnpcs.api.event.NPCRightClickEvent;
-import net.thenextlvl.service.api.npc.CharacterController;
-import net.thenextlvl.service.api.npc.event.CharacterDamageEvent;
-import net.thenextlvl.service.api.npc.event.EntityDamageCharacterEvent;
-import net.thenextlvl.service.api.npc.event.PlayerInteractCharacterEvent;
-import net.thenextlvl.service.api.npc.event.PlayerInteractCharacterEvent.InteractionType;
+import net.thenextlvl.service.api.character.CharacterController;
+import net.thenextlvl.service.api.character.event.CharacterDamageEvent;
+import net.thenextlvl.service.api.character.event.EntityDamageCharacterEvent;
+import net.thenextlvl.service.api.character.event.PlayerInteractCharacterEvent;
+import net.thenextlvl.service.api.character.event.PlayerInteractCharacterEvent.InteractionType;
 import net.thenextlvl.service.model.character.citizens.CitizensCharacter;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
