+++ plugin/src/main/java/net/thenextlvl/service/model/character/citizens/CitizensCharacter.java
@@ -7,7 +7,7 @@ import net.citizensnpcs.api.trait.trait.MobType;
 import net.citizensnpcs.api.trait.trait.PlayerFilter;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.MiniMessage;
-import net.thenextlvl.service.api.npc.Character;
+import net.thenextlvl.service.api.character.Character;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Server;
