+++ plugin/src/main/java/net/thenextlvl/service/model/character/fancy/FancyCharacter.java
@@ -4,7 +4,7 @@ import de.oliver.fancynpcs.api.FancyNpcsPlugin;
 import de.oliver.fancynpcs.api.Npc;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.MiniMessage;
-import net.thenextlvl.service.api.npc.Character;
+import net.thenextlvl.service.api.character.Character;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Server;
